pyproject.toml
[project]
name = "fastbrain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.40.0",
    "python-frontmatter>=1.1.0",
    "python-slugify>=8.0.0",
]
